[Types] Fix misleading filename; Split into separate files
