+++ src/types/Actions.ts
@@ -1,0 +1,10 @@
+export enum Action {
+  randomize = 'randomize',
+  next = 'next',
+  previous = 'previous',
+  restore = 'restore',
+  echo = 'echo',
+  save = 'save',
+  reset = 'reset',
+  toggleChar = 'toggleChar',
+}
